Fix bug where only 1 family per individual is saved.
This bug was due to the fact that FAM* records can occur at several
priorities. In the case of the GEDCOMs I'm working with, that's either 0
or 1. If 0, it only needs to be written once. If 1, it needs to be
written immediately and on every occurance. Now families show up
correctly on RootsWeb.